Refactor Runtime::Factories

+++ taichi/runtime/runtime.cpp
@@ -4,7 +4,7 @@ TLANG_NAMESPACE_BEGIN
 
 std::unique_ptr<Runtime> Runtime::create(Arch arch) {
   auto &factories = get_factories();
-  if (auto factory = factories.find((int)arch); factory != factories.end()) {
+  if (auto factory = factories.find(arch); factory != factories.end()) {
     return factory->second();
   } else {
     return nullptr;
